Fixes on size types

+++ include/atmi_int.h
@@ -766,7 +766,7 @@ extern NDRX_API int ndrx_load_new_env(char *file);
 extern NDRX_API int ndrx_generic_q_send(char *queue, char *data, long len, long flags, unsigned int msg_prio);
 extern NDRX_API int ndrx_generic_q_send_2(char *queue, char *data, long len, long flags, long tout, unsigned int msg_prio);
 extern NDRX_API int ndrx_generic_qfd_send(mqd_t q_descr, char *data, long len, long flags);
-extern NDRX_API long ndrx_generic_q_receive(mqd_t q_descr, char *q_str, 
+extern NDRX_API ssize_t ndrx_generic_q_receive(mqd_t q_descr, char *q_str, 
         struct mq_attr *reply_q_attr,
         char *buf, long buf_max, 
         unsigned *prio, long flags);
